+++ CHANGELOG.md
@@ -4,9 +4,19 @@
 
 ### Changed
 
+- Configuration - Typo fix in error-msg-not-authorised-redirect
+- Nuvlabox release - Published flag support
+- Module - New subtype applications set
+
+## [6.0.9] - 2023-01-06
+
+### Changed
+
 - Cloud-entry-point - Status should be 500 when ES is down
 - Email content test - Make current date constant for tests
-- Module - New subtype applications set
+- Notifications - Updated to use Kafka 3.x client. Updated the schema of the
+  configuration of subscriptions to notifications. Removed creation of
+  individual subscriptions.
 
 ## [6.0.8] - 2022-12-19